+++ lib/joystick/joystick_action.dart
@@ -181,6 +181,7 @@ class JoystickAction {
           event: ActionEvent.MOVE,
           intensity: intensity,
           radAngle: radAngle,
+          
         ),
       );
     } else {